+++ src/main/resources/db/changelog/changeset/V001__account-service_balance_DDL.sql
@@ -1,7 +1,7 @@
 CREATE TABLE balance(
     id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY UNIQUE,
-    authorization_balance BIGINT,
-    actual_balance BIGINT NOT NULL,
+    authorization_balance DOUBLE,
+    actual_balance DOUBLE NOT NULL,
     created_at TIMESTAMPTZ DEFAULT current_timestamp NOT NULL,
     updated_at TIMESTAMPTZ DEFAULT current_timestamp,
     version NUMERIC,
